env.requirements.txt
phpsh==1.3
python-magic==0.4.12
